requirements.txt
fastapi>=0.94.1
uvicorn>=0.21.1
pydantic>=1.10.6
email-validator==1.1.1
databases[postgresql]>=0.7.0
SQLAlchemy>=1.4.42
alembic>=1.10.2
h11==0.12.0
httptools==0.5.0
psycopg2-binary==2.8.4
starlette==0.26.1
typing==3.7.4.1
typing-extensions==4.2.0
websockets==10.4
asgiref==3.4.1
click==8.0.1
python-dotenv==0.18.0
python-dateutil
pytz